Update number and operator buttons
This commit adds functionality to the operator buttons so that when they
are clicked, their text content is added to the 'operator' variable,
which is used in the 'operate' function. Additionally, the number button
functionality has been updated so that when a number button is clicked,
its text content is added to either the 'firstNum' or 'secondNum'
variable, depending on whether an operator has been set yet.

The 'firstNum' and 'secondNum' values have been updated.