Update calculator display and logic for inputs
Added logic to replace the result in the main display with the next
digit(s) to be calculated if secondNum is not yet set.

Added code to display the clicked operator value in the minor display.

Move first number in the calculation and the first operator to the minor
display when operator is clicked the first time, whilst keeping firstNum
value in the main display.