Add logic to handle change of operator value
If the same operator is clicked multiple times, the value stays the
same in the operator variable.

The operator text in displayMinor will also not change.

If a different operator is clicked the operator variable will change to
that operator value. Then the operator text in displayMinor will be
replaced with the new operator text.